ajout d'un test sur MessageBundle
