--- src/MsgBundle/Tests/Controller/DefaultControllerTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace MsgBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class DefaultControllerTest extends WebTestCase
-{
-    public function testIndex()
-    {
-
-    }
-}
